pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dcsbpi"
version = "0.1.0"
description = "Raspberry Pi companion clients for DCSBiosRP2350"
authors = [ { name = "Your Name" } ]
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.9"
dependencies = [
    "pyserial>=3.5",
    "websockets>=10.0",
    "gpiozero>=1.6.2",
    "pyyaml>=6.0",
    "pillow>=9.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]
